Neural Network classifier to automate occurrence-based conservation assessments

iucnn() defines a neural network for predicting the conservation status of species given species-level predictors calculated from occurrence records. This is an implementation of the IUCNN model so it works in the tidymodels framework.

iucnn(
  mode = "classification",
  engine = "keras",
  layers = NULL,
  dropout = NULL,
  epochs = NULL
)

Arguments

mode

A single character string for the prediction outcome mode. Only "classification" is allowed.

engine

A single character string specifying the engine to use.

layers

A string specification of the hidden units in each layer, e.g. "40_20" for a two-layer network with a 40-unit layer then a 20-unit layer.

dropout

A number between 0 (inclusive) and 1 denoting the proportion of model parameters randomly set to zero during model training.

epochs

An integer for the number of training iterations.

Details

Currently only the binary threatened/not threatened classification is implemented.

Examples

parsnip::show_engines("iucnn")
#> # A tibble: 1 × 2
#>   engine mode          
#>   <chr>  <chr>         
#> 1 keras  classification

iucnn(layers="40_20", dropout=0.3, epochs=10)
#> Neural network assessments specification (classification)
#> 
#> Main Arguments:
#>   layers = 40_20
#>   dropout = 0.3
#>   epochs = 10
#>   learn_rate = 10
#> 
#> Computational engine: keras 
#>