Interface to a neural network model for automated species conservation assessments via keras

iucnn_model() builds a sequential keras model from a string specification of the units in each layer of the network. Regularisation is by dropout.

iucnn_model(
  x,
  y,
  layers = "30",
  dropout = 0,
  epochs = 30,
  validation_data = NULL,
  save_history = FALSE,
  ...
)

Arguments

x

A dataframe of matrix of predictors

y

A vector (factor or character) of outcome data.

layers

A string specification of the number of hidden units in each layer, e.g. "40_20"

dropout

The proportion of parameters to set to zero.

epochs

An integer for the number of passes through the data.

validation_data

A dataframe of data for validation with columns for predictiors in x and the outcome y, or a list of two matrices for the predictors and outcome.

save_history

Logical, whether to save the loss values for each epoch during training.

...

compilation and fitting arguments to pass to the keras model.

Value

A keras model object.