Rule based classification

rule_based() defines a model that uses fixed rules to classify an input.

rule_based(mode = "classification", engine = "eoo")

Arguments

mode

A single character string for the prediction outcome mode. Only "classification" is allowed.

engine

A single character string specifying the engine to use.

Details

As the model uses fixed rules, fitting it just creates a model object. Different sets of rules are implemented as model engines.

Information about the class of model fit can be found in rule_based_assessment. The engine-specific pages for the different rule sets can be found in:

eoo_threshold (default)

conr_thresholds

More information on how parsnip is used for modeling is at https://www.tidymodels.org/.

Examples

parsnip::show_engines("rule_based")
#> # A tibble: 2 × 2
#>   engine mode          
#>   <chr>  <chr>         
#> 1 eoo    classification
#> 2 conr   classification

rule_based(mode="classification", engine="conr")
#> Model Specification (classification)
#> 
#> Computational engine: conr 
#>