Extinction risk predictors for Myrcia species.

A dataset containing predictors to use in automated conservation methods for species of the genus Myrcia. These predictors were prepared for the paper Evidence-based guidelines for automated assessments of plant species, and described in detail there. The dataset contains a mixture of assessed and unassessed species. Assessments were compiled from the IUCN Red List v.2021-3 and predictors were calculated across observed localities for each species as downloaded from GBIF.

Usage

myrcia_predictors

Format

A data frame with 666 rows and 18 variables:

wcvp_id

a unique identifier for the species, from WCVP

,
eoo

Extent of occurrence (EOO), in km^2

,
human_footprint

a measure of the impact of human activities, in 2009

,
human_population

human population density in 2020, in persons/km^2

,
forest_loss

proportion of a species observed localities that underwent forest loss between 2000-2019

,
annual_temperature

mean average temperature, in degrees-C

,
precipitation_driest

precipitation in the driest quarter, in mm

,
elevation

elevation, in m

,
conr_eoo

EOO, as calculated by ConR, in km^2

,
conr_aoo

Area of occupancy (AOO), as calculated by ConR, in km^2

,
conr_locs

Number of locations estimated by ConR

,
conr_obs

Number of unique observations calculated by ConR

,
n_specimens

Total number of observations

,
centroid_latitude

Latitude of range centroid, in decimal-degrees

,
name

species name, as listed in WCVP

,
category

IUCN Red List category, if assessed

,
genus

genus of species, as listed in WCVP

,
section

section of species, as detailed in Lucas et al. (2019)

Source

https://doi.org/10.32942/osf.io/zxq6s